data/class refactor to remove snap items
