+++ lib/app/home/home_page.dart
@@ -136,11 +136,11 @@ class HomePageState extends ConsumerState<HomePage> {
     if (isKeeper) {
       context.goNamed(AppRoutes.circleCreate);
     } else {
-      final SnapCircle? createdCircle =
+      final Circle? createdCircle =
           await CircleCreateNonKeeper.showNonKeeperCreateDialog(context);
       if (mounted && createdCircle != null) {
         context.goNamed(AppRoutes.circle,
-            params: {'id': createdCircle.snapSession.id});
+            params: {'id': createdCircle.session.id});
       }
     }
     if (mounted) {
