+++ lib/models/index.dart
@@ -3,6 +3,7 @@ export 'active_session.dart';
 export 'audio_indicator.dart';
 export 'auth_user.dart';
 export 'circle.dart';
+export 'circle_template.dart';
 export 'circle_theme.dart';
 export 'communication_device.dart';
 export 'data_urls.dart';
@@ -15,8 +16,6 @@ export 'session.dart';
 export 'session_data_message.dart';
 export 'session_participant.dart';
 export 'session_token.dart';
-export 'snap_circle.dart';
-export 'snap_session.dart';
 export 'system_video.dart';
 export 'topic.dart';
 export 'user_auth_account_state.dart';
