+++ lib/models/session.dart
@@ -1,28 +1,35 @@
 import 'package:totem/models/index.dart';
 
-abstract class Session {
+class Session {
   late final String id;
-  late final SnapCircle circle;
+  late final Circle circle;
   late String topic;
+  Map<String, Participant> participants = {};
+  Map<String, Map<String, dynamic>> participantData = {};
 
-  String get ref {
-    return "";
+  SessionState get state {
+    return circle.state;
   }
 
-  SessionState get state;
-  set state(SessionState stateVal);
+  set state(SessionState stateVal) {
+    circle.state = stateVal;
+  }
+
+  int get participantCount {
+    return participantData.length;
+  }
+
+  String get ref {
+    return circle.ref;
+  }
 
   Session.fromJson(Map<String, dynamic> json,
       {required this.id, required this.circle}) {
     topic = json['topic'] ?? "";
-  }
-
-  Map<String, dynamic> toJson() {
-    Map<String, dynamic> item = {
-      "topic": topic,
-      "state": state.name,
-    };
-    return item;
+    if (json['participants'] != null) {
+      participantData =
+          Map<String, Map<String, dynamic>>.from(json['participants']);
+    }
   }
 
   @override
@@ -35,4 +42,19 @@ abstract class Session {
 
   @override
   int get hashCode => id.hashCode;
+
+  Map<String, dynamic> toJson({bool includeParticipants = true}) {
+    Map<String, dynamic> item = {
+      "topic": topic,
+      "state": state.name,
+    };
+    if (includeParticipants) {
+      Map<String, Map<String, dynamic>> partData = {};
+      for (var key in participants.keys) {
+        partData[key] = participants[key]!.toJson();
+      }
+      item["participants"] = partData;
+    }
+    return item;
+  }
 }
