--- lib/models/snap_session.dart
@@ -1,54 +1,0 @@
-import 'package:totem/models/index.dart';
-
-class SnapSession extends Session {
-  static const String snap = "snap";
-
-  Map<String, Participant> participants = {};
-  Map<String, Map<String, dynamic>> participantData = {};
-
-  int get participantCount {
-    return participantData.length;
-  }
-
-  SnapSession.fromJson(
-    Map<String, dynamic> json, {
-    required SnapCircle circle,
-  }) : super.fromJson(
-          json,
-          id: circle.id,
-          circle: circle,
-        ) {
-    if (json['participants'] != null) {
-      participantData =
-          Map<String, Map<String, dynamic>>.from(json['participants']);
-    }
-  }
-
-  @override
-  String get ref {
-    return circle.ref;
-  }
-
-  @override
-  SessionState get state {
-    return circle.state;
-  }
-
-  @override
-  set state(SessionState stateVal) {
-    circle.state = stateVal;
-  }
-
-  @override
-  Map<String, dynamic> toJson({bool includeParticipants = true}) {
-    Map<String, dynamic> data = super.toJson();
-    if (includeParticipants) {
-      Map<String, Map<String, dynamic>> partData = {};
-      for (var key in participants.keys) {
-        partData[key] = participants[key]!.toJson();
-      }
-      data["participants"] = partData;
-    }
-    return data;
-  }
-}
