+++ pubspec.yaml
@@ -55,7 +55,7 @@ dependencies:
   flutter_foreground_task: ^3.5.4
   flutter_markdown: ^0.6.12
   flutter_reorderable_list: ^1.2.0
-  flutter_riverpod: ^1.0.0
+  flutter_riverpod: ^2.1.0
   flutter_sim_country_code: ^0.1.2
   flutter_spinbox: ^0.10.0
   flutter_svg: ^1.0.3
@@ -70,7 +70,7 @@ dependencies:
   keyboard_actions: ^4.0.0
   libphonenumber_plugin: ^0.2.3
   logger: ^1.0.0-nullsafety.0
-  lucide_icons: ^0.92.0
+  lucide_icons: ^0.98.0
   noise_meter: ^3.1.0
   package_info_plus: ^3.0.0
   path_provider: ^2.0.10
